@@ -2,13 +2,10 @@ apply plugin: 'com.android.library'
 group = 'com.github.ybq'
 version = '1.0'
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 31
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 28
-        versionCode 1
-        versionName "1.0"
+        targetSdkVersion 31
     }
 }
 dependencies {
