@@ -1,20 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 31
     defaultConfig {
-        applicationId "com.safesoft.pro.distribute"
-        minSdkVersion 19
-        targetSdkVersion 30
-        versionCode 531
-        versionName "1.1.530"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        applicationId "com.safesoft.proapp.distribute"
+        minSdkVersion 24
+        targetSdkVersion 31
+        versionCode 537
+        versionName "1.2"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         multiDexEnabled true //important
-
-//        ndk     {
-//
-//            abiFilters "armeabi-v7a", "x86", "armeabi", "mips"
-//        }
 
     }
 
@@ -55,7 +50,7 @@ allprojects {
 }
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.4.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
     implementation('de.keyboardsurfer.android.widget:crouton:1.8.5@aar') {
@@ -65,17 +60,17 @@ dependencies {
 
     implementation 'com.haozhang.libary:android-slanted-textview:1.2'
     implementation 'com.haozhang.libary:android-slanted-textview:1.2'
-    implementation 'com.android.support:appcompat-v7:27.1.0'
-    implementation 'com.android.support:cardview-v7:27.1.0'
-    implementation 'com.android.support:design:27.1.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'com.google.android.material:material:1.5.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
     implementation 'com.amulyakhare:com.amulyakhare.textdrawable:1.0.1'
     implementation 'org.greenrobot:eventbus:3.0.0'
     implementation 'com.uniquestudio:checkbox:1.0.1'
     implementation 'com.labo.kaji:swipeawaydialog:0.1.0'
     implementation 'com.baoyz.swipemenulistview:library:1.3.0'
     implementation 'com.emmasuzuki:easyform:1.1.0'
-    implementation 'com.google.android.gms:play-services-maps:11.6.0'
+    implementation 'com.google.android.gms:play-services-maps:18.0.2'
     implementation 'com.github.paolorotolo:expandableheightlistview:1.0.0'
     implementation 'com.victor:lib:1.0.4'
     implementation 'eu.inloop:simplerecycleradapter:0.3.1'
@@ -84,11 +79,11 @@ dependencies {
     implementation 'com.weiwangcn.betterspinner:library-material:1.1.0'
     implementation 'com.jakewharton:butterknife:8.0.1'
     implementation 'com.dropbox.core:dropbox-core-sdk:3.0.5'
-    implementation 'com.squareup.okhttp3:okhttp:3.7.0'
+    implementation 'com.squareup.okhttp3:okhttp:3.10.0'
     implementation 'pl.coreorb:selection-dialogs:0.9.1'
     //Button
     implementation 'com.rilixtech:materialfancybuttons:1.8.7'
-    implementation 'com.google.android.gms:play-services-vision:11.6.0'
+    implementation 'com.google.android.gms:play-services-vision:20.1.3'
     implementation 'me.dm7.barcodescanner:core:1.9.8'
     implementation 'me.dm7.barcodescanner:zbar:1.9.8'
     //implementation 'me.dm7.barcodescanner:zxing:1.9.8'
@@ -116,7 +111,13 @@ dependencies {
     implementation files('libs/opencsv-3.9.jar')
     implementation files('libs/httpmime-4.0.3.jar')
     implementation files('libs/dropbox-android-sdk-1.6.3.jar')
-    testImplementation 'junit:junit:4.12'
+    testImplementation 'junit:junit:4.13.2'
     implementation files('libs/commons-net-1.4.1.jar')
+
+    implementation ("com.journeyapps:zxing-android-embedded:4.2.0"){
+        exclude group: 'com.google.zxing', module: 'android-core'
+        transitive = false
+    }
+
 
 }
