@@ -4,13 +4,10 @@ group = 'com.ramotion.foldingcell'
 version = '1.1.0'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 31
     defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 26
-        versionCode 3
-        versionName '1.1.0'
+        minSdkVersion 24
+        targetSdkVersion 31
     }
     buildTypes {
         release {
@@ -18,8 +15,6 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
         debug {
-            versionNameSuffix " Debug"
-            debuggable true
             testCoverageEnabled = true
         }
     }
