@@ -3,14 +3,11 @@ apply plugin: 'com.android.library'
 group='com.github.nekocode'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 31
 
     defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 28
-        versionCode 2
-        versionName "1.6.2"
+        minSdkVersion 24
+        targetSdkVersion 31
     }
     buildTypes {
         release {
@@ -22,5 +19,5 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:support-annotations:27.0.1'
+    implementation 'androidx.annotation:annotation:1.3.0'
 }
