@@ -3,7 +3,7 @@ package com.edwardvanraak.materialbarcodescanner;
 import android.app.Dialog;
 import android.hardware.Camera;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.view.WindowManager;
@@ -22,8 +22,6 @@ import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
 
 import java.io.IOException;
-
-import static junit.framework.Assert.assertNotNull;
 
 public class MaterialBarcodeScannerActivity extends AppCompatActivity {
 
@@ -71,7 +69,6 @@ public class MaterialBarcodeScannerActivity extends AppCompatActivity {
 
     private void setupLayout() {
         final TextView topTextView = (TextView) findViewById(R.id.topText);
-        assertNotNull(topTextView);
         String topText = mMaterialBarcodeScannerBuilder.getText();
         if(!mMaterialBarcodeScannerBuilder.getText().equals("")){
             topTextView.setText(topText);
@@ -103,7 +100,6 @@ public class MaterialBarcodeScannerActivity extends AppCompatActivity {
     private void setupButtons() {
         final LinearLayout flashOnButton = (LinearLayout)findViewById(R.id.flashIconButton);
         final ImageView flashToggleIcon = (ImageView)findViewById(R.id.flashIcon);
-        assertNotNull(flashOnButton);
         flashOnButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -228,7 +224,7 @@ public class MaterialBarcodeScannerActivity extends AppCompatActivity {
     private void clean() {
         EventBus.getDefault().removeStickyEvent(MaterialBarcodeScanner.class);
         if (mCameraSourcePreview != null) {
-            mCameraSourcePreview.release();
+           // mCameraSourcePreview.release();
             mCameraSourcePreview = null;
         }
         if(mSoundPoolPlayer != null){
