@@ -3,13 +3,11 @@ package com.edwardvanraak.materialbarcodescanner;
 import android.app.Activity;
 import android.graphics.Color;
 import android.hardware.Camera;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.ViewGroup;
 
 import com.google.android.gms.vision.barcode.Barcode;
 import com.google.android.gms.vision.barcode.BarcodeDetector;
-
-import org.greenrobot.eventbus.EventBus;
 
 public class MaterialBarcodeScannerBuilder {
 
